Renamed files regarding spoonancular api to make function clearer
