+++ spoonacular_logic/query_spoonacular.py
